chore: move LogContentProvider to own file
