+++ src/extension/src/providers/content/LogEntry.ts
